Fixes reader to properly read from entire file
Noticed that there were few ways main could know if there was another line to be read. Even if there is a blank line in between more lines. This implementation fixes this issue by recognizing '\n' lines as valid entries. Afterall, the last line will not contain a '\n' character.